Exercício_07 Média aritmética.
Adicionando "FLOAT" no lugar de "INT".

+++ media_aritimetica.py
@@ -1,4 +1,4 @@
-Nota_01 = int( input('Primeira nota do aluno: '))
-Nota_02 = int(input('Segunda nota do aluno: '))
+Nota_01 = float( input('Primeira nota do aluno: '))
+Nota_02 = float(input('Segunda nota do aluno: '))
 media = (Nota_01 + Nota_02) /2
 print('A média entre {:.2f} e {:.2f} é igual a {:.2f}'.format(Nota_01, Nota_02, media))
